improve input validation
