+++ kibana-reports/server/routes/utils/reportHelper.ts
@@ -13,22 +13,15 @@
  * permissions and limitations under the License.
  */
 
-import puppeteer from 'puppeteer';
-import createDOMPurify from 'dompurify';
-import { JSDOM } from 'jsdom';
 import {
   FORMAT,
   REPORT_TYPE,
   REPORT_STATE,
-  CONFIG_INDEX_NAME,
   LOCAL_HOST,
   DELIVERY_TYPE,
   EMAIL_FORMAT,
-  DEFAULT_REPORT_FOOTER,
-  DEFAULT_REPORT_HEADER,
 } from './constants';
-import { RequestParams } from '@elastic/elasticsearch';
-import { getFileName, callCluster } from './helpers';
+import { callCluster, updateToES, saveToES } from './helpers';
 import {
   ILegacyClusterClient,
   ILegacyScopedClusterClient,
@@ -37,123 +30,7 @@ import {
 import { createSavedSearchReport } from './savedSearchReportHelper';
 import { ReportSchemaType } from '../../model';
 import { CreateReportResultType } from './types';
-
-const window = new JSDOM('').window;
-const DOMPurify = createDOMPurify(window);
-
-export const createVisualReport = async (
-  reportParams: any,
-  queryUrl: string,
-  logger: Logger
-): Promise<CreateReportResultType> => {
-  const coreParams = reportParams.core_params;
-  // parse params
-  const reportSource = reportParams.report_source;
-  const reportName = reportParams.report_name;
-  const windowWidth = coreParams.window_width;
-  const windowHeight = coreParams.window_height;
-  const reportFormat = coreParams.report_format;
-
-  // TODO: polish default header, maybe add a logo, depends on UX design
-  const header = coreParams.header ? DOMPurify.sanitize(coreParams.header) : DEFAULT_REPORT_HEADER;
-  const footer = coreParams.footer ? DOMPurify.sanitize(coreParams.footer) : DEFAULT_REPORT_FOOTER;
-  // set up puppeteer
-  const browser = await puppeteer.launch({
-    headless: true,
-    /**
-     * TODO: temp fix to disable sandbox when launching chromium on Linux instance
-     * https://github.com/puppeteer/puppeteer/blob/main/docs/troubleshooting.md#setting-up-chrome-linux-sandbox
-     */
-    args: ['--no-sandbox', '--disable-setuid-sandbox'],
-  });
-  const page = await browser.newPage();
-  page.setDefaultNavigationTimeout(0);
-  page.setDefaultTimeout(60000); // use 60s timeout instead of default 30s
-  logger.info(`original queryUrl ${queryUrl}`);
-  await page.goto(queryUrl, { waitUntil: 'networkidle0' });
-  logger.info(`page url ${page.url()}`);
-  logger.info(`page url includes login? ${page.url().includes('login')}`);
-
-  /**
-   * TODO: This is a workaround to simulate a login to security enabled domain.
-   * Need better handle.
-   */
-  if (page.url().includes('login')) {
-    logger.info(
-      'domain enables security, redirecting to login page, start simulating login'
-    );
-    await page.type('[placeholder=Username]', 'admin', { delay: 30 });
-    await page.type('[placeholder=Password]', 'admin', { delay: 30 });
-    await page.click("[type='submit']");
-    await page.waitForNavigation();
-    logger.info(
-      `Done logging in, currently at page: ${page.url()} \nGo to queryUrl again`
-    );
-    await page.goto(queryUrl, { waitUntil: 'networkidle0' });
-    logger.info(`wait for network idle, the current page url: ${page.url()}`);
-  }
-
-  await page.setViewport({
-    width: windowWidth,
-    height: windowHeight,
-  });
-
-  let buffer;
-  let element;
-  // crop content
-  if (reportSource === REPORT_TYPE.dashboard) {
-    await page.waitForSelector('#dashboardViewport');
-    element = await page.$('#dashboardViewport');
-  } else if (reportSource === REPORT_TYPE.visualization) {
-    await page.waitForSelector('.visChart');
-    element = await page.$('.visChart');
-  }
-
-  const screenshot = await element.screenshot({ fullPage: false });
-
-  /**
-   * Sets the content of the page to have the header be above the trimmed screenshot
-   * and the footer be below it
-   */
-  // TODO: need to convert header from markdown to html, either do it on server side, or on client side.
-  // Email body conversion is done from client side
-  await page.setContent(`
-    <!DOCTYPE html>
-    <html>
-      <div>
-      ${header}
-        <img src="data:image/png;base64,${screenshot.toString('base64')}">
-      ${footer}
-      </div>
-    </html>
-    `);
-
-  // create pdf or png accordingly
-  if (reportFormat === FORMAT.pdf) {
-    const scrollHeight = await page.evaluate(
-      () => document.documentElement.scrollHeight
-    );
-
-    buffer = await page.pdf({
-      margin: undefined,
-      width: windowWidth,
-      height: scrollHeight + 'px',
-      printBackground: true,
-      pageRanges: '1',
-    });
-  } else if (reportFormat === FORMAT.png) {
-    buffer = await page.screenshot({
-      fullPage: true,
-    });
-  }
-
-  const curTime = new Date();
-  const timeCreated = curTime.valueOf();
-  const fileName = `${getFileName(reportName, curTime)}.${reportFormat}`;
-  await browser.close();
-
-  return { timeCreated, dataUrl: buffer.toString('base64'), fileName };
-};
+import { createVisualReport } from './visualReportHelper';
 
 export const createReport = async (
   isScheduledTask: boolean,
@@ -177,20 +54,8 @@ export const createReport = async (
   const reportParams = reportDefinition.report_params;
   const reportSource = reportParams.report_source;
 
-  // validate url
-  let queryUrl = report.query_url;
-  const { origin } = new URL(report.query_url);
-  /**
-   * Only URL in WHATWG format is accepted.
-   * e.g http://dasd@test.com/random?size=50 will be considered as invalid and throw error, due to the "@" symbol
-   * is not part of the url origin according to https://nodejs.org/api/url.html#url_url_strings_and_url_objects
-   */
-  if (report.query_url.search(origin) >= 0) {
-    queryUrl = report.query_url.replace(origin, LOCAL_HOST);
-  } else {
-    throw Error(`query url is not valid: ${queryUrl}`);
-  }
-
+  // compose url
+  const queryUrl = `${LOCAL_HOST}${report.query_url}`;
   try {
     // generate report
     if (reportSource === REPORT_TYPE.savedSearch) {
@@ -239,63 +104,6 @@ export const createReport = async (
   }
 
   return createReportResult;
-};
-
-export const saveToES = async (
-  isScheduledTask: boolean,
-  report: ReportSchemaType,
-  esClient: ILegacyClusterClient | ILegacyScopedClusterClient
-) => {
-  const timePending = Date.now();
-  const saveParams: RequestParams.Index = {
-    index: CONFIG_INDEX_NAME.report,
-    body: {
-      ...report,
-      state: REPORT_STATE.pending,
-      last_updated: timePending,
-      time_created: timePending,
-    },
-  };
-  const esResp = await callCluster(
-    esClient,
-    'index',
-    saveParams,
-    isScheduledTask
-  );
-
-  return esResp;
-};
-
-export const updateToES = async (
-  isScheduledTask: boolean,
-  reportId: string,
-  esClient: ILegacyClusterClient | ILegacyScopedClusterClient,
-  state: string,
-  createReportResult?: CreateReportResultType
-) => {
-  const timeStamp = createReportResult
-    ? createReportResult.timeCreated
-    : Date.now();
-  // update report document with state "created" or "error"
-  const updateParams: RequestParams.Update = {
-    id: reportId,
-    index: CONFIG_INDEX_NAME.report,
-    body: {
-      doc: {
-        state: state,
-        last_updated: timeStamp,
-        time_created: timeStamp,
-      },
-    },
-  };
-  const esResp = await callCluster(
-    esClient,
-    'update',
-    updateParams,
-    isScheduledTask
-  );
-
-  return esResp;
 };
 
 export const deliverReport = async (
