+++ kibana-reports/server/utils/validationHelper.ts
@@ -13,8 +13,24 @@
  * permissions and limitations under the License.
  */
 
-export interface CreateReportResultType {
-  timeCreated: number;
-  dataUrl: string;
-  fileName: string;
-}
+import { LOCAL_HOST } from '../routes/utils/constants';
+
+export const isValidRelativeUrl = (relativeUrl: string) => {
+  try {
+    new URL(`${LOCAL_HOST}${relativeUrl}`);
+  } catch (_) {
+    return false;
+  }
+  return true;
+};
+
+/**
+ * moment.js isValid() API fails to validate time duration, so use regex
+ * https://github.com/moment/moment/issues/1805
+ **/
+
+export const isValidDuration = (duration: string) => {
+  return !!duration.match(
+    /^(-?)P(?=\d|T\d)(?:(\d+)Y)?(?:(\d+)M)?(?:(\d+)([DW]))?(?:T(?:(\d+)H)?(?:(\d+)M)?(?:(\d+(?:\.\d+)?)S)?)?$/
+  );
+};
